feat: реализовать ТЗ диплома, этап #2 / разработка - v6.0

+++ ewm-service/src/main/java/ru/practicum/ewm/event/model/EventFullDto.java
@@ -24,5 +24,5 @@ public class EventFullDto {
     CategoryDto category;
     UserShortDto initiator;
     Long confirmedRequests;
-    Long views;
+    Integer views;
 }
